fix: added tests
