+++ src/test/java/browser/BrowserTest.java
@@ -1,0 +1,201 @@
+package browser;
+
+import static org.junit.Assert.*;
+
+import com.looksee.models.Browser;
+import com.looksee.models.BrowserConnectionHelper;
+import com.looksee.models.enums.BrowserEnvironment;
+import com.looksee.models.enums.BrowserType;
+import com.looksee.services.BrowserService;
+import java.awt.image.BufferedImage;
+import java.io.File;
+import java.io.IOException;
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.security.KeyManagementException;
+import java.security.NoSuchAlgorithmException;
+import java.util.Map;
+import javax.imageio.ImageIO;
+import org.jsoup.Jsoup;
+import org.jsoup.nodes.Document;
+import org.junit.jupiter.api.Test;
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebDriverException;
+import org.openqa.selenium.WebElement;
+
+
+/**
+ *
+ */
+public class BrowserTest {
+	
+	@Test
+	public void verifyCleanAttributeValuesString(){
+		String src_example = "PDF: Pearson\'s Watson-Glaser II Critical Thinking Appraisal and CPP\'s CPI 260 assessment";
+		String clean_src = BrowserService.cleanAttributeValues(src_example);
+		System.err.println("clean src: " +clean_src);
+		assertTrue("PDF: Pearson's Watson-Glaser II Critical Thinking Appraisal and CPP's CPI 260 assessment".equals(clean_src));
+		
+		src_example = "section[contains(@class,\"dashboard-content-wrapper \t     col-lg-10 col-lg-offset-1 \t\t    col-xs-12 col-xs-offset-0    bordered-box\")]";
+		clean_src = BrowserService.cleanAttributeValues(src_example);
+		System.err.println("clean src: " +clean_src);
+		assertTrue("section[contains(@class,\\\"dashboard-content-wrapper col-lg-10 col-lg-offset-1 col-xs-12 col-xs-offset-0 bordered-box\\\")]".equals(clean_src));
+	}
+
+	
+	@Test
+	public void verifyUrlReaderForHttps() throws MalformedURLException {
+		URL url = new URL("https://www.amazon.com");
+		try {
+			String output = Browser.URLReader(url);
+			System.out.println("output           :: "+output);
+			assertTrue(output!= null);
+			assertTrue(!output.isEmpty());
+		} catch (KeyManagementException | NoSuchAlgorithmException | IOException e) {
+			e.printStackTrace();
+			assertTrue(false);
+		}
+	}
+	
+	@Test
+	public void verifyGenerateParentXpath(){
+		System.out.println("testing 123");
+		try{
+			Browser browser = BrowserConnectionHelper.getConnection(BrowserType.FIREFOX, BrowserEnvironment.DISCOVERY);
+			browser.navigateTo("https://look-see.com");
+			WebElement element = browser.getDriver().findElement(By.xpath("//li//a[contains(@href,'features.html')]/../../.."));
+			Map<String, String> attributes = browser.extractAttributes(element);
+
+			BrowserService browser_service = new BrowserService();
+			String xpath = browser_service.generateXpath(element, browser.getDriver(), attributes);
+			System.err.println("XPATH :: " + xpath);
+			//log.info("clean src: " +clean_src);
+		//	Assert.assertTrue("concat('This is a embedded ', '\"', 'path', '\"', '')".equals(clean_src));
+		}
+		catch(Exception e){
+			e.printStackTrace();
+		}
+	}
+	
+	//@Test
+	//NOTE: Commented out because testbed url is not correct anymore
+	public void scrollToElementInChrome() throws MalformedURLException{
+		Browser browser = BrowserConnectionHelper.getConnection(BrowserType.CHROME, BrowserEnvironment.DISCOVERY);
+		browser.navigateTo("https://qa-testbed.qanairy.com/viewport_pages/element_out_of_view_y_axis.html");
+
+		String xpath = "//button";
+		WebElement element = browser.getDriver().findElement(By.xpath(xpath));
+		
+		browser.scrollToElement(element);
+		
+		assertEquals(0, browser.getXScrollOffset());
+		assertEquals(553, browser.getYScrollOffset());
+	}
+	
+	public void verifyAttributes() throws MalformedURLException{
+		int cnt = 0;
+		do{
+			try{
+				Browser browser = BrowserConnectionHelper.getConnection(BrowserType.FIREFOX, BrowserEnvironment.DISCOVERY);
+				browser.navigateTo("https://qa-testbed.qanairy.com/elements/index.html");
+				WebElement element = browser.getDriver().findElement(By.xpath("//button"));
+				
+				Map<String, String> attributes = browser.extractAttributes(element);
+				
+				assertTrue(attributes.containsKey("id"));
+				assertEquals(1, attributes.get("id").length());
+				
+				assertTrue(attributes.containsKey("class"));
+				assertEquals(3, attributes.get("class").length());
+				
+				assertTrue(attributes.containsKey("style"));
+				assertEquals(1, attributes.get("style").length());
+				break;
+			}
+			catch(WebDriverException e){
+				
+			}
+			cnt++;
+		}while(cnt<5);
+	}
+	
+	@Test
+	public void verifyCleanSrc() {
+		String src = "<html><script src=''></script><link href=''/><style>.style{}</style><head></head><body><div>This is a test</div></body></html>";
+		
+		String cleaned_src = Browser.cleanSrc(src);
+		assertFalse(cleaned_src.contains("<script"));
+		assertFalse(cleaned_src.contains("<style"));
+	}
+	
+	@Test
+	public void verifyCleanUrl() throws MalformedURLException {
+		String url1 = "https://look-see.com";
+		URL url = new URL(url1);
+		
+		System.err.println("url :: "+url.getHost());
+		
+	}
+	
+	//NOTE: This is used for manual testing to generate a full page screenshot.
+	//@Test
+	public void verifyFullPageScreenshot(){
+		System.out.println("testing 123");
+		try{
+			Browser browser = BrowserConnectionHelper.getConnection(BrowserType.CHROME, BrowserEnvironment.DISCOVERY);
+			browser.navigateTo("https://medium.com");
+			
+			BufferedImage scrn2 = browser.getFullPageScreenshotShutterbug();
+			ImageIO.write(scrn2, "png", new File("/home/deepthought/fullimage-shutterbug.png"));
+
+			//log.info("clean src: " +clean_src);
+		//	Assert.assertTrue("concat('This is a embedded ', '\"', 'path', '\"', '')".equals(clean_src));
+		}
+		catch(Exception e){
+			//e.printStackTrace();
+		}
+	}
+	
+	@Test
+	public void testGeneralizeSrc() {
+		String expected_result = "<html><head></head><body><app-footer></app-footer></body></html>";
+		String src = "<app-footer _ngcontent-wiq-c74=\"\" class=\"footer\" _nghost-wiq-c23=\"\"></app-footer>  ";
+		String generalized_src = BrowserService.generalizeSrc(src);
+		System.out.println(generalized_src);
+		assertTrue(expected_result.equals(generalized_src));
+	}
+	
+	@Test
+	public void testCommentRemoval() {
+		String expected_result = "<html>\n"
+				+ " <head></head>\n"
+				+ " <body>\n"
+				+ "  <app-footer _ngcontent-wiq-c74=\"\" class=\"footer\" _nghost-wiq-c23=\"\">\n"
+				+ "   <div></div>\n"
+				+ "  </app-footer> \n"
+				+ " </body>\n"
+				+ "</html>";
+		
+		String src = "<app-footer _ngcontent-wiq-c74=\"\" class=\"footer\" _nghost-wiq-c23=\"\"><div></div><!-- this is a comment --></app-footer>  ";
+		Document html_doc = Jsoup.parse(src);
+
+		BrowserService.removeComments(html_doc);
+		
+		assert(expected_result.equals(html_doc.html()));
+	}
+
+	@Test
+	public void testCleanSrc(){
+		String expected_result = "<html> <head></head> <body> <app-footer _ngcontent-wiq-c74=\"\" class=\"footer\" _nghost-wiq-c23=\"\"></app-footer>  </body></html>";
+		String src = "<app-footer _ngcontent-wiq-c74=\"\" class=\"footer\" _nghost-wiq-c23=\"\"></app-footer>  ";
+		String generalized_src = Browser.cleanSrc(src);
+		System.out.println(generalized_src);
+		assertTrue(expected_result.equals(generalized_src));
+
+		String expected_result2 = "<html> <head></head> <body> <app-footer _ngcontent-wiq-c74=\"\" class=\"footer\" _nghost-wiq-c23=\"\"></app-footer>  </body></html>";
+		String src2 = "<html lang=\"en\"> <head> <link rel=\"stylesheet\" type=\"text/css\" href=\"https://cdn.unstack.com/static/css/aos.v2.min.css\"> <link rel=\"stylesheet\" type=\"text/css\" href=\"https://cdn.unstack.com/static/css/highlight.min.css\">  <meta name=\"viewport\" content=\"width=device-width, initial-scale=1, maximum-scale=1.0\">  <title>Automated UX audits | Look-see</title>  <meta property=\"og:title\" content=\"Automated UX audits | Look-see\">  <meta name=\"twitter:title\" content=\"Automated UX audits | Look-see\">  <meta property=\"og:type\" content=\"website\">  <meta name=\"twitter:card\" content=\"summary_large_image\">  <meta name=\"description\" content=\"Look-see performs automated UX audits and helps you understand where to make improvements\">  <meta property=\"og:description\" content=\"Look-see performs automated UX audits and helps you understand where to make improvements\">  <meta name=\"twitter:description\" content=\"Look-see performs automated UX audits and helps you understand where to make improvements\">  <meta name=\"keywords\" content=\"Look-see, UX, user experience, user experiences, UX audit, website, webpage, webpage evaluation, webpage audit, website audit, information architecture, aesthetics, visual design, content, written content, blog, readability\">  <link rel=\"canonical\" href=\"https://www.look-see.com/\">  <meta property=\"og:url\" content=\"https://www.look-see.com/\">  <meta property=\"og:image\" content=\"https://cdn.spark.app/media/looksee/image/demo_notebook.png\">  <meta name=\"twitter:image\" content=\"https://cdn.spark.app/media/looksee/image/demo_notebook.png\">  <link rel=\"icon\" type=\"image/png\" href=\"https://cdn.spark.app/media/looksee/icon/red.png\">  <link rel=\"stylesheet\" type=\"text/css\" href=\"https://cdn.unstack.com/static/css/main.css?v=fa6b83fe\">  <link rel=\"stylesheet\" type=\"text/css\" href=\"https://cdn.unstack.com/static/css/plugins/plugins.css?v=fa6b83fe\">  <link rel=\"preconnect\" href=\"https://fonts.gstatic.com\">  <link rel=\"stylesheet\" type=\"text/css\" href=\"https://fonts.googleapis.com/css?family=Open Sans:300,400,500,700|Muli:300,400,500,700&amp;display=block\">  <!-- GA -->   <!-- /GA -->  <link rel=\"stylesheet\" href=\"https://cdn.spark.app/media/site/css/looksee-5143295b4cd141013104f0f48c696434b0600cc1135732660bfe320ffc067ee7.css\">  </head>  <body id=\"unstack\" class=\"sriracha site_page item-f0c26be2-550e-4f61-a855-47c1fbb400e2\" data-item-type=\"site_page\" data-static-path=\"https://cdn.unstack.com/static/\" data-path-prefix=\"\" data-unstack-sid=\"\" data-currency-disambiguation=\"\" data-domain=\"https://www.look-see.com/\" data-product-cache-check-delay=\"5\" data-aos-easing=\"cubic-bezier(0.165, 0.84, 0.44, 1)\" data-aos-duration=\"750\" data-aos-delay=\"0\">  <div class=\"header-container\">   <!-- Header -->   <header class=\" primary-header solid light-bg header-menu-layout-right box-padding-top-4 box-padding-bottom-4 position-fixed  \">   <div class=\"container\">    <button id=\"mobile-nav-toggle\" class=\"hamburger hamburger--spin\" type=\"button\"> <span class=\"hamburger-box\"><span class=\"hamburger-inner\"></span></span> </button>    <div class=\"logo\">    <a href=\"/\" id=\"header_logo_link\" data-section=\"header\" data-event=\"click\" data-category=\"header_logo\" data-action=\"click\" data-label=\"Look-see\"> <img src=\"https://cdn.spark.app/media/looksee/image/wordmark_black.png\" alt=\"Look-see\"> </a>    </div>    <div class=\"nav-group empty-cart\">    <nav id=\"primary-nav\">     <ul>     <li> <a class=\"link active\" href=\"https://www.look-see.com/\" data-event=\"click\" data-category=\"footer_nav\" data-action=\"click\" data-section=\"footer\">Home</a> </li>     <li> <a class=\"link \" href=\"https://www.look-see.com/product\" data-event=\"click\" data-category=\"footer_nav\" data-action=\"click\" data-section=\"footer\">Product</a>      <svg class=\"dropdown-icon\" xmlns=\"http://www.w3.org/2000/svg\" viewbox=\"0 0 20 20\">      <path fill=\"currentColor\" d=\"M7.05 9.293L6.343 10 12 15.657l1.414-1.414L9.172 10l4.242-4.243L12 4.343z\"></path>      </svg>      <div class=\"mobile-dropdown-icon\">      <svg xmlns=\"http://www.w3.org/2000/svg\" viewbox=\"0 0 20 20\">       <path fill=\"currentColor\" d=\"M7.05 9.293L6.343 10 12 15.657l1.414-1.414L9.172 10l4.242-4.243L12 4.343z\"></path>      </svg>      </div>      <div class=\"secondary-navigation\">      <ul>       <li> <a class=\"\" href=\"https://www.look-see.com/product\" data-event=\"click\" data-category=\"header_nav\" data-action=\"click\" data-section=\"header\">Product</a> </li>       <li> <a class=\"\" href=\"https://www.look-see.com/technology\" data-event=\"click\" data-category=\"header_nav\" data-action=\"click\" data-section=\"header\">Technology</a> </li>      </ul>      </div> </li>     <li> <a class=\"link \">Use cases</a>      <svg class=\"dropdown-icon\" xmlns=\"http://www.w3.org/2000/svg\" viewbox=\"0 0 20 20\">      <path fill=\"currentColor\" d=\"M7.05 9.293L6.343 10 12 15.657l1.414-1.414L9.172 10l4.242-4.243L12 4.343z\"></path>      </svg>      <div class=\"mobile-dropdown-icon\">      <svg xmlns=\"http://www.w3.org/2000/svg\" viewbox=\"0 0 20 20\">       <path fill=\"currentColor\" d=\"M7.05 9.293L6.343 10 12 15.657l1.414-1.414L9.172 10l4.242-4.243L12 4.343z\"></path>      </svg>      </div>      <div class=\"secondary-navigation\">      <ul>       <li> <a class=\"\" href=\"https://www.look-see.com/founders\" data-event=\"click\" data-category=\"header_nav\" data-action=\"click\" data-section=\"header\">Founder</a> </li>       <li> <a class=\"\" href=\"https://www.look-see.com/agencies\" data-event=\"click\" data-category=\"header_nav\" data-action=\"click\" data-section=\"header\">Agency</a> </li>       <li> <a class=\"\" href=\"https://www.look-see.com/marketers\" data-event=\"click\" data-category=\"header_nav\" data-action=\"click\" data-section=\"header\">Marketer</a> </li>      </ul>      </div> </li>     <li> <a class=\"link \">Pricing</a>      <svg class=\"dropdown-icon\" xmlns=\"http://www.w3.org/2000/svg\" viewbox=\"0 0 20 20\">      <path fill=\"currentColor\" d=\"M7.05 9.293L6.343 10 12 15.657l1.414-1.414L9.172 10l4.242-4.243L12 4.343z\"></path>      </svg>      <div class=\"mobile-dropdown-icon\">      <svg xmlns=\"http://www.w3.org/2000/svg\" viewbox=\"0 0 20 20\">       <path fill=\"currentColor\" d=\"M7.05 9.293L6.343 10 12 15.657l1.414-1.414L9.172 10l4.242-4.243L12 4.343z\"></path>      </svg>      </div>      <div class=\"secondary-navigation\">      <ul>       <li> <a class=\"\" href=\"https://www.look-see.com/pricing/business\" data-event=\"click\" data-category=\"header_nav\" data-action=\"click\" data-section=\"header\">Business</a> </li>       <li> <a class=\"\" href=\"https://www.look-see.com/pricing/agency\" data-event=\"click\" data-category=\"header_nav\" data-action=\"click\" data-section=\"header\">Agency</a> </li>      </ul>      </div> </li>     <li> <a class=\"link \" href=\"https://www.look-see.com/blog\" data-event=\"click\" data-category=\"footer_nav\" data-action=\"click\" data-section=\"footer\">Blog</a> </li>     <!--=== AUTHENTICATED-USER-NAV ===--->     <li><a class=\"button button-primary \" href=\"https://app.look-see.com\">Try it for Free</a></li>     </ul>    </nav>    <div class=\"header-icons\">    </div>    </div>   </div>   </header>   <!-- /Header -->  </div>  <!-- Header -->  <header class=\" primary-header solid light-bg header-menu-layout-right box-padding-top-4 box-padding-bottom-4  substrate \">   <div class=\"container\">   <button id=\"mobile-nav-toggle\" class=\"hamburger hamburger--spin\" type=\"button\"> <span class=\"hamburger-box\"><span class=\"hamburger-inner\"></span></span> </button>   <div class=\"logo\">    <a href=\"/\" id=\"header_logo_link\" data-section=\"header\" data-event=\"click\" data-category=\"header_logo\" data-action=\"click\" data-label=\"Look-see\"> <img src=\"https://cdn.spark.app/media/looksee/image/wordmark_black.png\" alt=\"Look-see\"> </a>   </div>   <div class=\"nav-group empty-cart\">    <nav id=\"primary-nav\">    <ul>     <li> <a class=\"link active\" href=\"https://www.look-see.com/\" data-event=\"click\" data-category=\"footer_nav\" data-action=\"click\" data-section=\"footer\">Home</a> </li>     <li> <a class=\"link \" href=\"https://www.look-see.com/product\" data-event=\"click\" data-category=\"footer_nav\" data-action=\"click\" data-section=\"footer\">Product</a>     <svg class=\"dropdown-icon\" xmlns=\"http://www.w3.org/2000/svg\" viewbox=\"0 0 20 20\">      <path fill=\"currentColor\" d=\"M7.05 9.293L6.343 10 12 15.657l1.414-1.414L9.172 10l4.242-4.243L12 4.343z\"></path>     </svg>     <div class=\"mobile-dropdown-icon\">      <svg xmlns=\"http://www.w3.org/2000/svg\" viewbox=\"0 0 20 20\">      <path fill=\"currentColor\" d=\"M7.05 9.293L6.343 10 12 15.657l1.414-1.414L9.172 10l4.242-4.243L12 4.343z\"></path>      </svg>     </div>     <div class=\"secondary-navigation\">      <ul>      <li> <a class=\"\" href=\"https://www.look-see.com/product\" data-event=\"click\" data-category=\"header_nav\" data-action=\"click\" data-section=\"header\">Product</a> </li>      <li> <a class=\"\" href=\"https://www.look-see.com/technology\" data-event=\"click\" data-category=\"header_nav\" data-action=\"click\" data-section=\"header\">Technology</a> </li>      </ul>     </div> </li>     <li> <a class=\"link \">Use cases</a>     <svg class=\"dropdown-icon\" xmlns=\"http://www.w3.org/2000/svg\" viewbox=\"0 0 20 20\">      <path fill=\"currentColor\" d=\"M7.05 9.293L6.343 10 12 15.657l1.414-1.414L9.172 10l4.242-4.243L12 4.343z\"></path>     </svg>     <div class=\"mobile-dropdown-icon\">      <svg xmlns=\"http://www.w3.org/2000/svg\" viewbox=\"0 0 20 20\">      <path fill=\"currentColor\" d=\"M7.05 9.293L6.343 10 12 15.657l1.414-1.414L9.172 10l4.242-4.243L12 4.343z\"></path>      </svg>     </div>     <div class=\"secondary-navigation\">      <ul>      <li> <a class=\"\" href=\"https://www.look-see.com/founders\" data-event=\"click\" data-category=\"header_nav\" data-action=\"click\" data-section=\"header\">Founder</a> </li>      <li> <a class=\"\" href=\"https://www.look-see.com/agencies\" data-event=\"click\" data-category=\"header_nav\" data-action=\"click\" data-section=\"header\">Agency</a> </li>      <li> <a class=\"\" href=\"https://www.look-see.com/marketers\" data-event=\"click\" data-category=\"header_nav\" data-action=\"click\" data-section=\"header\">Marketer</a> </li>      </ul>     </div> </li>     <li> <a class=\"link \">Pricing</a>     <svg class=\"dropdown-icon\" xmlns=\"http://www.w3.org/2000/svg\" viewbox=\"0 0 20 20\">      <path fill=\"currentColor\" d=\"M7.05 9.293L6.343 10 12 15.657l1.414-1.414L9.172 10l4.242-4.243L12 4.343z\"></path>     </svg>     <div class=\"mobile-dropdown-icon\">      <svg xmlns=\"http://www.w3.org/2000/svg\" viewbox=\"0 0 20 20\">      <path fill=\"currentColor\" d=\"M7.05 9.293L6.343 10 12 15.657l1.414-1.414L9.172 10l4.242-4.243L12 4.343z\"></path>      </svg>     </div>     <div class=\"secondary-navigation\">      <ul>      <li> <a class=\"\" href=\"https://www.look-see.com/pricing/business\" data-event=\"click\" data-category=\"header_nav\" data-action=\"click\" data-section=\"header\">Business</a> </li>      <li> <a class=\"\" href=\"https://www.look-see.com/pricing/agency\" data-event=\"click\" data-category=\"header_nav\" data-action=\"click\" data-section=\"header\">Agency</a> </li>      </ul>     </div> </li>     <li> <a class=\"link \" href=\"https://www.look-see.com/blog\" data-event=\"click\" data-category=\"footer_nav\" data-action=\"click\" data-section=\"footer\">Blog</a> </li>     <!--=== AUTHENTICATED-USER-NAV ===--->     <li><a class=\"button button-primary \" href=\"https://app.look-see.com\">Try it for Free</a></li>    </ul>    </nav>    <div class=\"header-icons\">    </div>   </div>   </div>  </header>  <!-- /Header -->  <link rel=\"stylesheet\" href=\"https://cdn.spark.app/media/page/css/f13c279a7d43d3d917d3d1358da54c5b.css\">  <!-- Component --> <section data-id=\"89fc77cb-c88a-4e41-9e49-a66d986b8e0d\" class=\"section component-1ca0a679-fc4a-4ba7-a1e6-b1fd9bec611e section-89fc77cb-c88a-4e41-9e49-a66d986b8e0d dark-bg height-normal right normal ratio-1-1 mb-show-after\" data-property-layout=\"right\" data-property-style=\"normal\" data-property-ratio=\"ratio-1-1\" data-property-gap=\"4\" data-property-mobile=\"mb-show-after\" data-property-addons=\"none\" data-addons=\"none\">   <div class=\"section-background section-background-media section-background-image\">   </div>   <div class=\"section-background\" style=\"background-color: rgba(255,0,81,0)\">  &nbsp;  </div>  <div class=\"box box-1380beafba0a491692f0d16ff11ce8a6 section-box\" style=\"padding-top: 6rem;padding-bottom: 6rem;\">  <div class=\"box-content\" style=\"margin-right: auto;margin-left: auto;max-width: 1500px;\">    <div class=\"media-container\" style=\"gap:4rem\">    <div class=\"box vertical-align middle box-d71f41772a854081845bb59ac16d095b media-box\" style=\"padding-top: 0rem;padding-right: 0rem;padding-bottom: 0rem;padding-left: 0rem;\">    <div class=\"box-background\">    <div class=\"box-background shadow-none\" style=\"background-color: None; opacity: 1; border-color: #000; border-width: 0px; border-style: none; border-radius: 0px; box-shadow: none\">     &nbsp;    </div>    </div>    <div class=\"box-content\" style=\"margin-right: auto;margin-left: auto;max-width: 800px;\">     <div class=\"unstack-media\">      <picture class=\"responsive-media\" style=\"padding-bottom: 54.52054794520548%;\">      <source data-srcset=\"https://cdn.spark.app/media/looksee/image/demo_notebook.png.w535.png 1.0x\" type=\"image/png\" media=\"(max-width: 575px)\" srcset=\"https://cdn.spark.app/media/looksee/image/demo_notebook.png.w535.png 1.0x\">     <source data-srcset=\"https://cdn.spark.app/media/looksee/image/demo_notebook.png.w727.png 1.0x\" type=\"image/png\" media=\"(max-width: 767px)\" srcset=\"https://cdn.spark.app/media/looksee/image/demo_notebook.png.w727.png 1.0x\">     <source data-srcset=\"https://cdn.spark.app/media/looksee/image/demo_notebook.png.w456.png 1.0x\" type=\"image/png\" media=\"(max-width: 991px)\" srcset=\"https://cdn.spark.app/media/looksee/image/demo_notebook.png.w456.png 1.0x\">     <source data-srcset=\"https://cdn.spark.app/media/looksee/image/demo_notebook.png.w560.png 1.0x\" type=\"image/png\" media=\"(max-width: 1199px)\" srcset=\"https://cdn.spark.app/media/looksee/image/demo_notebook.png.w560.png 1.0x\">     <source data-srcset=\"https://cdn.spark.app/media/looksee/image/demo_notebook.png.w730.png 1.0x\" type=\"image/png\" srcset=\"https://cdn.spark.app/media/looksee/image/demo_notebook.png.w730.png 1.0x\">      <img data-src=\"https://cdn.spark.app/media/looksee/image/demo_notebook.png.w730.png\" alt=\"Look-see dashboard rendered on a laptop that shows introduction text describing that a 77% usability score is close to a delightful experience.\" data-id=\"media:04ff478d-be3e-4876-924e-485341f27ef0\" data-index=\"1-1-1\" width=\"730\" height=\"398\" src=\"https://cdn.spark.app/media/looksee/image/demo_notebook.png.w730.png\">      <noscript>       <img src=\"https://cdn.spark.app/media/looksee/image/demo_notebook.png.w730.png\" alt=\"Look-see dashboard rendered on a laptop that shows introduction text describing that a 77% usability score is close to a delightful experience.\" data-id=\"media:04ff478d-be3e-4876-924e-485341f27ef0\" data-index=\"1-1-1\" width=\"730\" height=\"398\">      </noscript>      </picture>     </div>     </div>   </div>    <div class=\"text-container\">     <div class=\"box vertical-align middle box-dbc55399b89a4f259214bd36b847cc12 text-box\" style=\"padding-top: 9rem;padding-right: 0rem;padding-left: 0rem;\">    <div class=\"box-background\">     <div class=\"box-background shadow-none\" style=\"background-color: None; opacity: 1; border-color: #000; border-width: 0px; border-style: none; border-radius: 0px; box-shadow: none\">     &nbsp;     </div>    </div>    <div class=\"box-content text-center\" style=\"margin-right: auto;margin-left: auto;max-width: 550px;\">      <div class=\"rich-text \">     <h1><span style=\"font-family: Open Sans\">Automated Accessibility Audits </span></h1>     <h3><span style=\"font-family: Muli\">Review your website in minutes and learn where you can improve your user experiences</span></h3>     </div>      <div class=\"main-actions\">      <div class=\"actions\">       <a id=\"89fc77cb-c88a-4e41-9e49-a66d986b8e0d-secondary-1-5-1-13-1\" class=\"button button-secondary\" data-event=\"click\" data-section=\"89fc77cb-c88a-4e41-9e49-a66d986b8e0d\" data-category=\"secondary_action\" data-action=\"click\" data-label=\"Get Started for Free\" data-value=\"1-5-1-13-1\" href=\"https://app.look-see.com\">Get Started for Free</a>     </div>      </div>     </div>    </div>    </div>    </div>   </div>  </div> </section> <!-- /Component -->  <!-- Component --> <section data-id=\"b898c736-6cd9-431f-b449-cd02d0ca89d2\" class=\"section component-1ca0a679-fc4a-4ba7-a1e6-b1fd9bec611e section-b898c736-6cd9-431f-b449-cd02d0ca89d2 light-bg height-normal left normal ratio-1-1 mb-show-after\" data-property-layout=\"left\" data-property-style=\"normal\" data-property-ratio=\"ratio-1-1\" data-property-gap=\"4\" data-property-mobile=\"mb-show-after\" data-property-addons=\"none\" data-addons=\"none\">  <div class=\"section-background\" style=\"background-color: rgba(255,255,255,1)\">  &nbsp;  </div>  <div class=\"box box-f713fd55f49e468b87e9e714b8c784b5 section-box\">  <div class=\"box-content\" style=\"margin-right: auto;margin-left: auto;max-width: 1200px;\">    <div class=\"media-container\" style=\"gap:4rem\">    <div class=\"box vertical-align middle box-d0512943456b42ebb31b849226a670af media-box\" style=\"padding-top: 0rem;padding-right: 0rem;padding-bottom: 0rem;padding-left: 0rem;\">    <div class=\"box-background\">    <div class=\"box-background shadow-none\" style=\"background-color: None; opacity: 1; border-color: #000; border-width: 0px; border-style: none; border-radius: 0px; box-shadow: none\">     &nbsp;    </div>    </div>    <div class=\"box-content aos-hide aos-init aos-animate\" style=\"margin-right: auto;margin-left: auto;max-width: 400px;\" data-aos=\"fade-up\">     <div class=\"unstack-media\">      <picture class=\"responsive-media\" style=\"padding-bottom: 138.75%;\">      <source data-srcset=\"https://cdn.spark.app/media/looksee/image/mobile_demo_2.png.w400.webp 1.0x\" type=\"image/webp\" srcset=\"https://cdn.spark.app/media/looksee/image/mobile_demo_2.png.w400.webp 1.0x\">     <source data-srcset=\"https://cdn.spark.app/media/looksee/image/mobile_demo_2.png.w400.png 1.0x\" type=\"image/png\" srcset=\"https://cdn.spark.app/media/looksee/image/mobile_demo_2.png.w400.png 1.0x\">      <img data-src=\"https://cdn.spark.app/media/looksee/image/mobile_demo_2.png.w400.png\" alt=\"Mobile screenshot of areas of improvement screen\" data-id=\"media:ed4931e5-f1e1-4b65-ab54-818829ac8eb5\" data-index=\"1-1-1\" width=\"400\" height=\"555\" src=\"https://cdn.spark.app/media/looksee/image/mobile_demo_2.png.w400.png\">      <noscript>       <img src=\"https://cdn.spark.app/media/looksee/image/mobile_demo_2.png.w400.png\" alt=\"Mobile screenshot of areas of improvement screen\" data-id=\"media:ed4931e5-f1e1-4b65-ab54-818829ac8eb5\" data-index=\"1-1-1\" width=\"400\" height=\"555\">      </noscript>      </picture>     </div>     </div>   </div>    <div class=\"text-container\">     <div class=\"box vertical-align middle box-7ecfc6cd45004ed6a3b062aef2ac3066 text-box\" style=\"padding-top: 8rem;padding-right: 0rem;padding-bottom: 0rem;padding-left: 0rem;\">    <div class=\"box-background\">     <div class=\"box-background shadow-none\" style=\"background-color: None; opacity: 1; border-color: #000; border-width: 0px; border-style: none; border-radius: 0px; box-shadow: none\">     &nbsp;     </div>    </div>    <div class=\"box-content text-left\" style=\"margin-right: auto;margin-left: auto;max-width: 400px;\">      <div class=\"rich-text \">     <h2 id=\"identify_core_improvement_areas\"><strong><span style=\"font-family: Open Sans\">Identify core improvement areas…</span></strong></h2>     <p><span style=\"font-family: Muli\">Look-see gives you recommendations to improve your website accessibility. We’ll tell you what parts of your website need improvement and provide recommendations to be compliant with section 508 of the ADA.</span></p>     <p><span style=\"font-family: Muli\">We’re here to help you create more accessible experiences for your users.</span></p>     <p><br></p>     </div>      <div class=\"main-actions\">      </div>     </div>    </div>    </div>    </div>   </div>  </div> </section> <!-- /Component -->  <!-- Component --> <section data-id=\"766a1c14-d324-4943-80b4-5a1c050d6b31\" class=\"section component-1ca0a679-fc4a-4ba7-a1e6-b1fd9bec611e section-766a1c14-d324-4943-80b4-5a1c050d6b31 light-bg height-normal right normal ratio-1-1 mb-show-after\" data-property-layout=\"right\" data-property-style=\"normal\" data-property-ratio=\"ratio-1-1\" data-property-gap=\"4\" data-property-mobile=\"mb-show-after\" data-property-addons=\"none\" data-addons=\"none\">  <div class=\"section-background\" style=\"background-color: rgba(235,248,255,1)\">  &nbsp;  </div>  <div class=\"box box-b8166085328d4628a28d1972936640a2 section-box\">  <div class=\"box-content\" style=\"margin-right: auto;margin-left: auto;max-width: 1200px;\">    <div class=\"media-container\" style=\"gap:4rem\">    <div class=\"box vertical-align middle box-a0008224a4964a918337e64f57865f8a media-box\" style=\"padding-top: 0rem;padding-right: 0rem;padding-bottom: 0rem;padding-left: 0rem;\">    <div class=\"box-background\">    <div class=\"box-background shadow-none\" style=\"background-color: None; opacity: 1; border-color: #000; border-width: 0px; border-style: none; border-radius: 0px; box-shadow: none\">     &nbsp;    </div>    </div>    <div class=\"box-content aos-hide aos-init\" style=\"margin-right: auto;margin-left: auto;max-width: 400px;\" data-aos=\"fade-up\">     <div class=\"unstack-media\">      <picture class=\"responsive-media\" style=\"padding-bottom: 138.75%;\">      <source data-srcset=\"https://cdn.spark.app/media/looksee/image/mobile_demo_1.png.w400.png 1.0x\" type=\"image/png\" srcset=\"https://cdn.spark.app/media/looksee/image/mobile_demo_1.png.w400.png 1.0x\">      <img data-src=\"https://cdn.spark.app/media/looksee/image/mobile_demo_1.png.w400.png\" alt=\"Mobile screenshot of score overview screen\" data-id=\"media:7ad28e18-1005-42af-8c44-c16a56c481b8\" data-index=\"1-1-1\" width=\"400\" height=\"555\" src=\"https://cdn.spark.app/media/looksee/image/mobile_demo_1.png.w400.png\">      <noscript>       <img src=\"https://cdn.spark.app/media/looksee/image/mobile_demo_1.png.w400.png\" alt=\"Mobile screenshot of score overview screen\" data-id=\"media:7ad28e18-1005-42af-8c44-c16a56c481b8\" data-index=\"1-1-1\" width=\"400\" height=\"555\">      </noscript>      </picture>     </div>     </div>   </div>    <div class=\"text-container\">     <div class=\"box vertical-align middle box-d3bca59e4fe043828e2310b786875412 text-box\" style=\"padding-top: 2rem;padding-right: 2rem;padding-left: 2rem;\">    <div class=\"box-background\">     <div class=\"box-background shadow-none\" style=\"background-color: None; opacity: 1; border-color: #000; border-width: 0px; border-style: none; border-radius: 0px; box-shadow: none\">     &nbsp;     </div>    </div>    <div class=\"box-content text-left\" style=\"margin-right: auto;margin-left: auto;max-width: 400px;\">      <div class=\"rich-text \">     <br>     <h2 id=\"wcag_compliance\"><strong><span style=\"font-family: Open Sans\">WCAG compliance</span></strong></h2>     <p><span style=\"font-family: Muli\">Our audits check for compliance with WCAG 2.1 guidelines so that you can easily ensure that you're providing accessible experiences. </span></p>     <p><span style=\"font-family: Muli\">We understand that your target audience isn't the same as every other company and that your accessibility needs may differ from other companies. That's why we put you in control! You can configure your WCAG compliance level to get results tailored to your needs. </span></p>     <p><br></p>     </div>      <div class=\"main-actions\">      </div>     </div>    </div>    </div>    </div>   </div>  </div> </section> <!-- /Component -->  <!-- Component --> <section data-id=\"d28a2b2b-9611-4210-9d39-c16f5e212b3c\" class=\"section component-1ca0a679-fc4a-4ba7-a1e6-b1fd9bec611e section-d28a2b2b-9611-4210-9d39-c16f5e212b3c light-bg height-normal left normal ratio-1-1 mb-show-after\" data-property-layout=\"left\" data-property-style=\"normal\" data-property-ratio=\"ratio-1-1\" data-property-gap=\"4\" data-property-mobile=\"mb-show-after\" data-property-addons=\"none\" data-addons=\"none\">  <div class=\"section-background\" style=\"background-color: rgba(255,255,255,1)\">  &nbsp;  </div>  <div class=\"box box-219d6352e06c40b282481351facf481b section-box\">  <div class=\"box-content\" style=\"margin-right: auto;margin-left: auto;max-width: 1200px;\">    <div class=\"media-container\" style=\"gap:4rem\">    <div class=\"box vertical-align middle box-c470a278cd1a45449b6abfc396999050 media-box\" style=\"padding-top: 0rem;padding-right: 0rem;padding-bottom: 0rem;padding-left: 0rem;\">    <div class=\"box-background\">    <div class=\"box-background shadow-none\" style=\"background-color: None; opacity: 1; border-color: #000; border-width: 0px; border-style: none; border-radius: 0px; box-shadow: none\">     &nbsp;    </div>    </div>    <div class=\"box-content aos-hide aos-init\" style=\"margin-right: auto;margin-left: auto;max-width: 400px;\" data-aos=\"fade-up\">     <div class=\"unstack-media\">      <picture class=\"responsive-media\" style=\"padding-bottom: 138.75%;\">      <source data-srcset=\"https://cdn.spark.app/media/looksee/image/mobile_demo_1.png.w400.png 1.0x\" type=\"image/png\" srcset=\"https://cdn.spark.app/media/looksee/image/mobile_demo_1.png.w400.png 1.0x\">      <img data-src=\"https://cdn.spark.app/media/looksee/image/mobile_demo_1.png.w400.png\" alt=\"Mobile screenshot of score overview screen\" data-id=\"media:7ad28e18-1005-42af-8c44-c16a56c481b8\" data-index=\"1-1-1\" width=\"400\" height=\"555\" src=\"https://cdn.spark.app/media/looksee/image/mobile_demo_1.png.w400.png\">      <noscript>       <img src=\"https://cdn.spark.app/media/looksee/image/mobile_demo_1.png.w400.png\" alt=\"Mobile screenshot of score overview screen\" data-id=\"media:7ad28e18-1005-42af-8c44-c16a56c481b8\" data-index=\"1-1-1\" width=\"400\" height=\"555\">      </noscript>      </picture>     </div>     </div>   </div>    <div class=\"text-container\">     <div class=\"box vertical-align middle box-5a0af9ebe59846cfa5490a817edd1eec text-box\" style=\"padding-top: 2rem;padding-right: 2rem;padding-left: 2rem;\">    <div class=\"box-background\">     <div class=\"box-background shadow-none\" style=\"background-color: None; opacity: 1; border-color: #000; border-width: 0px; border-style: none; border-radius: 0px; box-shadow: none\">     &nbsp;     </div>    </div>    <div class=\"box-content text-left\" style=\"margin-right: auto;margin-left: auto;max-width: 400px;\">      <div class=\"rich-text \">     <br>     <h2 id=\"learn_how_your_website_stacks_up\"><strong><span style=\"font-family: Open Sans\">Learn how your website stacks up…</span></strong></h2>     <p><span style=\"font-family: Muli\">We benchmark your website against your competitors to help you identify areas where you can stand out from the crowd. With our competitive analysis you can determine where providing better accessibility can give you a competitive edge.</span></p>     <p><span style=\"font-family: Muli\">Look-see will tell you how your website ranks and how you can improve accessibility for your customers.</span></p>     <p><br></p>     </div>      <div class=\"main-actions\">      </div>     </div>    </div>    </div>    </div>   </div>  </div> </section> <!-- /Component -->  <!-- Component --> <section data-id=\"d7920d4b-0699-45cd-8b33-40069704361b\" class=\"section component-1ca0a679-fc4a-4ba7-a1e6-b1fd9bec611e section-d7920d4b-0699-45cd-8b33-40069704361b light-bg height-normal right normal ratio-1-1 mb-show-after\" data-property-layout=\"right\" data-property-style=\"normal\" data-property-ratio=\"ratio-1-1\" data-property-gap=\"4\" data-property-mobile=\"mb-show-after\" data-property-addons=\"none\" data-addons=\"none\">  <div class=\"section-background\" style=\"background-color: rgba(235,248,255,1)\">  &nbsp;  </div>  <div class=\"box box-af70385ce3774c5aa18d771102fa0221 section-box\">  <div class=\"box-content\" style=\"margin-right: auto;margin-left: auto;max-width: 1200px;\">    <div class=\"media-container\" style=\"gap:4rem\">    <div class=\"box vertical-align middle box-c26b2b46f65c44ec8e9df3a1dba71e79 media-box\" style=\"padding-top: 0rem;padding-right: 0rem;padding-bottom: 0rem;padding-left: 0rem;\">    <div class=\"box-background\">    <div class=\"box-background shadow-none\" style=\"background-color: None; opacity: 1; border-color: #000; border-width: 0px; border-style: none; border-radius: 0px; box-shadow: none\">     &nbsp;    </div>    </div>    <div class=\"box-content aos-hide aos-init\" style=\"margin-right: auto;margin-left: auto;max-width: 350px;\" data-aos=\"fade-up\">     <div class=\"unstack-media\">      <picture class=\"responsive-media\" style=\"padding-bottom: 141.42857142857144%;\">      <source data-srcset=\"https://cdn.spark.app/media/looksee/image/report_cover.png.w350.webp 1.0x\" type=\"image/webp\" srcset=\"https://cdn.spark.app/media/looksee/image/report_cover.png.w350.webp 1.0x\">     <source data-srcset=\"https://cdn.spark.app/media/looksee/image/report_cover.png.w350.png 1.0x\" type=\"image/png\" srcset=\"https://cdn.spark.app/media/looksee/image/report_cover.png.w350.png 1.0x\">      <img data-src=\"https://cdn.spark.app/media/looksee/image/report_cover.png.w350.png\" alt=\"Look-see UX audit report cover screenshot\" data-id=\"media:354f82b3-223e-4aae-b878-ad663212bed4\" data-index=\"1-1-1\" width=\"350\" height=\"495\" src=\"https://cdn.spark.app/media/looksee/image/report_cover.png.w350.png\">      <noscript>       <img src=\"https://cdn.spark.app/media/looksee/image/report_cover.png.w350.png\" alt=\"Look-see UX audit report cover screenshot\" data-id=\"media:354f82b3-223e-4aae-b878-ad663212bed4\" data-index=\"1-1-1\" width=\"350\" height=\"495\">      </noscript>      </picture>     </div>     </div>   </div>    <div class=\"text-container\">     <div class=\"box vertical-align middle box-9855e9eef4af4d5e8dfd7961d06267da text-box\" style=\"padding-top: 10rem;padding-right: 0rem;padding-left: 0rem;\">    <div class=\"box-background\">     <div class=\"box-background shadow-none\" style=\"background-color: None; opacity: 1; border-color: #000; border-width: 0px; border-style: none; border-radius: 0px; box-shadow: none\">     &nbsp;     </div>    </div>    <div class=\"box-content text-left\" style=\"margin-right: auto;margin-left: auto;max-width: 400px;\">      <div class=\"rich-text \">     <h2 id=\"share_insights_with_stakeholders\"><span style=\"font-family: Open Sans\">Share insights with stakeholders</span></h2>     <p><span style=\"font-family: Muli\">Create reports that explain why various accessibility improvements matter. Our reports provide actionable insights that help stakeholders understand how changes may impact key performance metrics. </span></p>     <p id=\"\"><span style=\"font-family: Muli\">Don't just take our word for it. Check out our </span><span style=\"font-family: Muli\"><a href=\"https://storage.googleapis.com/look-see-inc-assets/sample-report.pdf\">sample report</a></span><span style=\"font-family: Muli\">.</span></p>     </div>      <div class=\"main-actions\">      </div>     </div>    </div>    </div>    </div>   </div>  </div> </section> <!-- /Component -->  <!-- Component --> <section data-id=\"71662dea-6774-4c85-a1b6-526ff68d5782\" class=\"section component-587345c8-2c48-4455-b180-3d83b23ee47a section-71662dea-6774-4c85-a1b6-526ff68d5782 light-bg height-normal center none xl c3 gap3\" data-property-layout=\"center\" data-property-style=\"none\" data-property-size=\"xl\" data-property-columns=\"c3\" data-property-gap=\"gap3\">  <div class=\"section-background\" style=\"background-color: rgba(247,248,251,1)\">  &nbsp;  </div>  <div class=\"box box-a258c9805f6a4b26ab87ebedf4b15eaa section-box\">  <div class=\"box-content\" style=\"margin-right: auto;margin-left: auto;max-width: 1100px;\">    <div class=\"box box-ffb0b546d97640ceac8ed678b547cf4e section-header-box\" style=\"padding-top: 0rem;padding-right: 0rem;padding-bottom: 6rem;padding-left: 0rem;\">   <div class=\"box-content text-left\" style=\"margin-right: auto;margin-left: auto;\">     <div class=\"rich-text\">    <h2 id=\"we_help_create_a_greater_impact\">We help create a greater impact…</h2>    <h3>The impact for you is so much more.</h3>    </div>   </div>   </div>    <div class=\"callouts\">   <div class=\"callout\">     <div class=\"unstack-media\">     <picture class=\"responsive-media\" style=\"padding-bottom: 106.25%;\">      <source data-srcset=\"https://cdn.spark.app/media/looksee/icon/lightbulb.png.w80.webp 1.0x\" type=\"image/webp\" srcset=\"https://cdn.spark.app/media/looksee/icon/lightbulb.png.w80.webp 1.0x\">     <source data-srcset=\"https://cdn.spark.app/media/looksee/icon/lightbulb.png.w80.png 1.0x\" type=\"image/png\" srcset=\"https://cdn.spark.app/media/looksee/icon/lightbulb.png.w80.png 1.0x\">      <img data-src=\"https://cdn.spark.app/media/looksee/icon/lightbulb.png.w80.png\" alt=\"lightbulb icon\" data-id=\"media:36769cd0-3fb0-41cb-b5b4-2fef4401e323\" data-index=\"1-0-1\" width=\"80\" height=\"85\" src=\"https://cdn.spark.app/media/looksee/icon/lightbulb.png.w80.png\">      <noscript>      <img src=\"https://cdn.spark.app/media/looksee/icon/lightbulb.png.w80.png\" alt=\"lightbulb icon\" data-id=\"media:36769cd0-3fb0-41cb-b5b4-2fef4401e323\" data-index=\"1-0-1\" width=\"80\" height=\"85\">      </noscript>     </picture>     </div>     <div class=\"rich-text \">    <h3>Raise Awareness</h3>    </div>    </div>   <div class=\"callout\">     <div class=\"unstack-media\">     <picture class=\"responsive-media\" style=\"padding-bottom: 100.0%;\">      <source data-srcset=\"https://cdn.spark.app/media/looksee/icon/smiley_face_secondary.png.w80.webp 1.0x\" type=\"image/webp\" srcset=\"https://cdn.spark.app/media/looksee/icon/smiley_face_secondary.png.w80.webp 1.0x\">     <source data-srcset=\"https://cdn.spark.app/media/looksee/icon/smiley_face_secondary.png.w80.png 1.0x\" type=\"image/png\" srcset=\"https://cdn.spark.app/media/looksee/icon/smiley_face_secondary.png.w80.png 1.0x\">      <img data-src=\"https://cdn.spark.app/media/looksee/icon/smiley_face_secondary.png.w80.png\" alt=\"smiley face icon\" data-id=\"media:2d521196-c37b-403a-b89f-018121c31522\" data-index=\"1-1-1\" width=\"80\" height=\"80\" src=\"https://cdn.spark.app/media/looksee/icon/smiley_face_secondary.png.w80.png\">      <noscript>      <img src=\"https://cdn.spark.app/media/looksee/icon/smiley_face_secondary.png.w80.png\" alt=\"smiley face icon\" data-id=\"media:2d521196-c37b-403a-b89f-018121c31522\" data-index=\"1-1-1\" width=\"80\" height=\"80\">      </noscript>     </picture>     </div>     <div class=\"rich-text \">    <h3>Increase Retention</h3>    </div>    </div>   <div class=\"callout\">     <div class=\"unstack-media\">     <picture class=\"responsive-media\" style=\"padding-bottom: 85.0%;\">      <source data-srcset=\"https://cdn.spark.app/media/looksee/icon/bounce.png.w80.webp 1.0x\" type=\"image/webp\" srcset=\"https://cdn.spark.app/media/looksee/icon/bounce.png.w80.webp 1.0x\">     <source data-srcset=\"https://cdn.spark.app/media/looksee/icon/bounce.png.w80.png 1.0x\" type=\"image/png\" srcset=\"https://cdn.spark.app/media/looksee/icon/bounce.png.w80.png 1.0x\">      <img data-src=\"https://cdn.spark.app/media/looksee/icon/bounce.png.w80.png\" alt=\"Bounce icon\" data-id=\"media:b47ac216-af5b-45dc-ac51-6c1d13eeb31e\" data-index=\"1-2-1\" width=\"80\" height=\"68\" src=\"https://cdn.spark.app/media/looksee/icon/bounce.png.w80.png\">      <noscript>      <img src=\"https://cdn.spark.app/media/looksee/icon/bounce.png.w80.png\" alt=\"Bounce icon\" data-id=\"media:b47ac216-af5b-45dc-ac51-6c1d13eeb31e\" data-index=\"1-2-1\" width=\"80\" height=\"68\">      </noscript>     </picture>     </div>     <div class=\"rich-text \">    <h3>Lower Bounce Rate</h3>    </div>    </div>   <div class=\"callout\">     <div class=\"unstack-media\">     <picture class=\"responsive-media\" style=\"padding-bottom: 98.75%;\">      <source data-srcset=\"https://cdn.spark.app/media/looksee/icon/increasing_graph.png.w80.webp 1.0x\" type=\"image/webp\" srcset=\"https://cdn.spark.app/media/looksee/icon/increasing_graph.png.w80.webp 1.0x\">     <source data-srcset=\"https://cdn.spark.app/media/looksee/icon/increasing_graph.png.w80.png 1.0x\" type=\"image/png\" srcset=\"https://cdn.spark.app/media/looksee/icon/increasing_graph.png.w80.png 1.0x\">      <img data-src=\"https://cdn.spark.app/media/looksee/icon/increasing_graph.png.w80.png\" alt=\"Graph with arrow point up and to the right\" data-id=\"media:9edd55b4-a8a5-4a15-8068-3ef68d95bc2c\" data-index=\"1-3-1\" width=\"80\" height=\"79\" src=\"https://cdn.spark.app/media/looksee/icon/increasing_graph.png.w80.png\">      <noscript>      <img src=\"https://cdn.spark.app/media/looksee/icon/increasing_graph.png.w80.png\" alt=\"Graph with arrow point up and to the right\" data-id=\"media:9edd55b4-a8a5-4a15-8068-3ef68d95bc2c\" data-index=\"1-3-1\" width=\"80\" height=\"79\">      </noscript>     </picture>     </div>     <div class=\"rich-text \">    <h3>Improve Engagement</h3>    </div>    </div>   <div class=\"callout\">     <div class=\"unstack-media\">     <picture class=\"responsive-media\" style=\"padding-bottom: 93.75%;\">      <source data-srcset=\"https://cdn.spark.app/media/looksee/icon/star.png.w80.webp 1.0x\" type=\"image/webp\" srcset=\"https://cdn.spark.app/media/looksee/icon/star.png.w80.webp 1.0x\">     <source data-srcset=\"https://cdn.spark.app/media/looksee/icon/star.png.w80.png 1.0x\" type=\"image/png\" srcset=\"https://cdn.spark.app/media/looksee/icon/star.png.w80.png 1.0x\">      <img data-src=\"https://cdn.spark.app/media/looksee/icon/star.png.w80.png\" alt=\"Star icon\" data-id=\"media:87064315-498a-4c53-b85c-f3fea56a2266\" data-index=\"1-4-1\" width=\"80\" height=\"75\" src=\"https://cdn.spark.app/media/looksee/icon/star.png.w80.png\">      <noscript>      <img src=\"https://cdn.spark.app/media/looksee/icon/star.png.w80.png\" alt=\"Star icon\" data-id=\"media:87064315-498a-4c53-b85c-f3fea56a2266\" data-index=\"1-4-1\" width=\"80\" height=\"75\">      </noscript>     </picture>     </div>     <div class=\"rich-text \">    <h3>Boost Credibility</h3>    </div>    </div>   <div class=\"callout\">     <div class=\"unstack-media\">     <picture class=\"responsive-media\" style=\"padding-bottom: 101.25%;\">      <source data-srcset=\"https://cdn.spark.app/media/looksee/icon/rocket_blastoff_j9SOWc2.png.w80.webp 1.0x\" type=\"image/webp\" srcset=\"https://cdn.spark.app/media/looksee/icon/rocket_blastoff_j9SOWc2.png.w80.webp 1.0x\">     <source data-srcset=\"https://cdn.spark.app/media/looksee/icon/rocket_blastoff_j9SOWc2.png.w80.png 1.0x\" type=\"image/png\" srcset=\"https://cdn.spark.app/media/looksee/icon/rocket_blastoff_j9SOWc2.png.w80.png 1.0x\">      <img data-src=\"https://cdn.spark.app/media/looksee/icon/rocket_blastoff_j9SOWc2.png.w80.png\" alt=\"Rocket taking off\" data-id=\"media:2ae6fcd5-7e32-4fff-af01-54e510dc1514\" data-index=\"1-5-1\" width=\"80\" height=\"81\" src=\"https://cdn.spark.app/media/looksee/icon/rocket_blastoff_j9SOWc2.png.w80.png\">      <noscript>      <img src=\"https://cdn.spark.app/media/looksee/icon/rocket_blastoff_j9SOWc2.png.w80.png\" alt=\"Rocket taking off\" data-id=\"media:2ae6fcd5-7e32-4fff-af01-54e510dc1514\" data-index=\"1-5-1\" width=\"80\" height=\"81\">      </noscript>     </picture>     </div>     <div class=\"rich-text \">    <h3>Better Accessibility</h3>    </div>    </div>   </div>   </div>  </div> </section> <!-- /Component -->  <!-- Component --> <section data-id=\"80742133-4a59-465d-9da7-0ca5081709a5\" class=\"section component-ba3e2b91-31dc-492f-9be1-19fcf32c4051 section-80742133-4a59-465d-9da7-0ca5081709a5 media-section light-bg height-normal\">  <div class=\"section-background\" style=\"background-color: rgba(235,248,255,1)\">  &nbsp;  </div>  <div class=\"box box-0ac662fc35794be381a406b8fe888f7e media-section section-box\">  <div class=\"box-content\" style=\"margin-right: auto;margin-left: auto;max-width: 1200px;\">    <div class=\"box box-f6892b2b0c744a088509f5b404dbd2c0 section-header-box\" style=\"padding-top: 0rem;padding-right: 0rem;padding-bottom: 2rem;padding-left: 0rem;\">   <div class=\"box-content text-center\" style=\"margin-right: auto;margin-left: auto;max-width: 400px;\">     <div class=\"rich-text\">    <h2 id=\"integrates_with_jira_trello_asana_google_drive_and_many_more\">Integrates with Jira, Trello, Asana, Google Drive, and many more</h2>    <p><span style=\"color: #3e547c\">Accessibility tasks are a natural part of your workflow with our integrations </span></p>    </div>   </div>   </div>    <div class=\"box box-2803af56bd6c44d1be1b648226ef80cc\" style=\"padding-top: 0rem;padding-right: 0rem;padding-bottom: 0rem;padding-left: 0rem;\">   <div class=\"box-background\">    <div class=\"box-background shadow-none\" style=\"background-color: None; opacity: 1; border-color: #000; border-width: 0px; border-style: none; border-radius: 0px; box-shadow: none\">    &nbsp;    </div>   </div>   <div class=\"box-content\" style=\"margin-right: auto;margin-left: auto;max-width: 800px;\">     <div class=\"unstack-media\">     <picture class=\"responsive-media\" style=\"padding-bottom: 42.5%;\">      <source data-srcset=\"https://cdn.spark.app/media/looksee/image/integration_logos.png.w535.webp 1.0x\" type=\"image/webp\" media=\"(max-width: 575px)\" srcset=\"https://cdn.spark.app/media/looksee/image/integration_logos.png.w535.webp 1.0x\">     <source data-srcset=\"https://cdn.spark.app/media/looksee/image/integration_logos.png.w535.png 1.0x\" type=\"image/png\" media=\"(max-width: 575px)\" srcset=\"https://cdn.spark.app/media/looksee/image/integration_logos.png.w535.png 1.0x\">     <source data-srcset=\"https://cdn.spark.app/media/looksee/image/integration_logos.png.w727.webp 1.0x\" type=\"image/webp\" media=\"(max-width: 767px)\" srcset=\"https://cdn.spark.app/media/looksee/image/integration_logos.png.w727.webp 1.0x\">     <source data-srcset=\"https://cdn.spark.app/media/looksee/image/integration_logos.png.w727.png 1.0x\" type=\"image/png\" media=\"(max-width: 767px)\" srcset=\"https://cdn.spark.app/media/looksee/image/integration_logos.png.w727.png 1.0x\">     <source data-srcset=\"https://cdn.spark.app/media/looksee/image/integration_logos.png.w800.webp 1.0x\" type=\"image/webp\" srcset=\"https://cdn.spark.app/media/looksee/image/integration_logos.png.w800.webp 1.0x\">     <source data-srcset=\"https://cdn.spark.app/media/looksee/image/integration_logos.png.w800.png 1.0x\" type=\"image/png\" srcset=\"https://cdn.spark.app/media/looksee/image/integration_logos.png.w800.png 1.0x\">      <img data-src=\"https://cdn.spark.app/media/looksee/image/integration_logos.png.w800.png\" alt=\"Jira, Trello, Asana, and Google Drive logos for integrations\" data-id=\"media:70aced84-c261-4983-9fce-a1473c3d2bbc\" data-index=\"1-1\" width=\"800\" height=\"340\" src=\"https://cdn.spark.app/media/looksee/image/integration_logos.png.w800.png\">      <noscript>      <img src=\"https://cdn.spark.app/media/looksee/image/integration_logos.png.w800.png\" alt=\"Jira, Trello, Asana, and Google Drive logos for integrations\" data-id=\"media:70aced84-c261-4983-9fce-a1473c3d2bbc\" data-index=\"1-1\" width=\"800\" height=\"340\">      </noscript>     </picture>     </div>    </div>   </div>   </div>  </div> </section> <!-- /Component -->  <!-- Component --> <section data-id=\"a9182d3e-7237-4ace-8428-c9cec72e212d\" class=\"section component-a54f1f46-fc91-414a-8e23-409c37671beb section-a9182d3e-7237-4ace-8428-c9cec72e212d dark-bg height-normal\">  <div class=\"section-background\" style=\"background-color: rgba(255,0,80,1)\">  &nbsp;  </div>  <div class=\"box box-4d527a4d4a0044be968bd081e9f9bcf6 section-box\">  <div class=\"box-content\" style=\"margin-right: auto;margin-left: auto;max-width: 800px;\">    <div class=\"box box-231c928dac9040028df71f992b8b2d46\">   <div class=\"box-background\">    <div class=\"box-background shadow-none\" style=\"background-color: None; opacity: 1; border-color: #000; border-width: 0px; border-style: none; border-radius: 0px; box-shadow: none\">    &nbsp;    </div>   </div>   <div class=\"box-content text-center\" style=\"margin-right: auto;margin-left: auto;max-width: 800px;\">     <div class=\"rich-text \">    <p><strong><span style=\"font-size: 4rem\">Stay on trend. Stay on brand.</span></strong></p>    <p><strong><span style=\"font-size: 4rem\">Stay ahead with Look-see.</span></strong></p>    </div>     <div class=\"actions\">     <a id=\"a9182d3e-7237-4ace-8428-c9cec72e212d-secondary-1-3\" class=\"button button-secondary\" data-event=\"click\" data-section=\"a9182d3e-7237-4ace-8428-c9cec72e212d\" data-category=\"secondary_action\" data-action=\"click\" data-label=\"Try it for free\" data-value=\"1-3\" href=\"https://app.look-see.com\">Try it for free</a>    </div>    </div>   </div>   </div>  </div> </section> <!-- /Component -->  <div class=\"sticky-bottom-spacer\"></div>  <footer class=\"main-footer dark-bg\">   <section class=\"primary-footer\">   <div class=\"container\">    <div class=\"info\">    <a href=\"/\" id=\"footer_logo_link\" class=\"logo\" data-section=\"footer\" data-event=\"click\" data-category=\"footer_logo\" data-action=\"click\" data-label=\"Look-see\"> <img src=\"https://cdn.spark.app/media/looksee/image/wordmark_red.png\" alt=\"Look-see\"> </a>    <div class=\"contact-info\">     <p></p>    <p>Boston, MA</p>    <p></p>    </div>    </div>    <div class=\"menus\">    <div class=\"menu\">     <h4>Product</h4>     <ul>     <li><a class=\"\" href=\"https://www.look-see.com/product\" data-event=\"click\" data-category=\"footer_nav\" data-action=\"click\" data-section=\"footer\">Product</a></li>     <li><a class=\"\" href=\"https://www.look-see.com/technology\" data-event=\"click\" data-category=\"footer_nav\" data-action=\"click\" data-section=\"footer\">Technology</a></li>     </ul>    </div>    <div class=\"menu\">     <h4>Use cases</h4>     <ul>     <li><a class=\"\" href=\"https://www.look-see.com/agencies\" data-event=\"click\" data-category=\"footer_nav\" data-action=\"click\" data-section=\"footer\">Agency</a></li>     <li><a class=\"\" href=\"https://www.look-see.com/founders\" data-event=\"click\" data-category=\"footer_nav\" data-action=\"click\" data-section=\"footer\">Founder</a></li>     <li><a class=\"\" href=\"https://www.look-see.com/marketers\" data-event=\"click\" data-category=\"footer_nav\" data-action=\"click\" data-section=\"footer\">Marketer</a></li>     </ul>    </div>    <div class=\"menu\">     <h4>Pricing</h4>     <ul>     <li><a class=\"\" href=\"https://www.look-see.com/pricing/business\" data-event=\"click\" data-category=\"footer_nav\" data-action=\"click\" data-section=\"footer\">Business</a></li>     <li><a class=\"\" href=\"https://www.look-see.com/pricing/agency\" data-event=\"click\" data-category=\"footer_nav\" data-action=\"click\" data-section=\"footer\">Agency</a></li>     </ul>    </div>    <div class=\"menu\">     <h4>Company</h4>     <ul>     <li><a class=\"active\" href=\"https://www.look-see.com/\" data-event=\"click\" data-category=\"footer_nav\" data-action=\"click\" data-section=\"footer\">Home</a></li>     <li><a class=\"\" href=\"https://www.look-see.com/blog\" data-event=\"click\" data-category=\"footer_nav\" data-action=\"click\" data-section=\"footer\">Blog</a></li>     <li><a class=\"\" href=\"https://www.look-see.com/about-us\" data-event=\"click\" data-category=\"footer_nav\" data-action=\"click\" data-section=\"footer\">About Us</a></li>     <li><a class=\"\" href=\"https://www.look-see.com/contact\" data-event=\"click\" data-category=\"footer_nav\" data-action=\"click\" data-section=\"footer\">Contact Us</a></li>     </ul>    </div>    </div>   </div>   </section>   <section class=\"secondary-footer\">   <div class=\"container\">    <!-- Copyright -->    <div style=\"text-align: center\" class=\"copyright\">    © Look-see    </div>    <!-- /Copyright -->    <div class=\"social-icons\">    <ul>     <li><a href=\"https://twitter.com/\" data-event=\"click\" data-category=\"social_link\" data-action=\"click\" data-section=\"footer\">     <svg width=\"24\" height=\"24\" viewbox=\"0 0 24 24\" fill=\"none\" xmlns=\"http://www.w3.org/2000/svg\">      <path d=\"M21.479 6.82729C21.4935 7.03153 21.4935 7.23576 21.4935 7.44188C21.4935 13.7224 16.5774 20.9656 7.58807 20.9656V20.9619C4.93258 20.9656 2.33226 20.2259 0.0967751 18.8311C0.482904 18.8762 0.870969 18.8988 1.26 18.8998C3.46065 18.9016 5.59839 18.1835 7.32968 16.8612C5.23839 16.8226 3.40452 15.4965 2.76387 13.5605C3.49645 13.6979 4.25129 13.6696 4.97032 13.4786C2.69032 13.0306 1.05 11.0824 1.05 8.81976C1.05 8.79906 1.05 8.77929 1.05 8.75953C1.72936 9.12753 2.49 9.33176 3.26807 9.35435C1.12065 7.95859 0.458711 5.18024 1.75548 3.008C4.23677 5.97741 7.89774 7.78259 11.8277 7.97365C11.4339 6.32282 11.9719 4.59294 13.2416 3.43247C15.21 1.63294 18.3058 1.72518 20.1561 3.63859C21.2506 3.42871 22.2997 3.03812 23.2597 2.48471C22.8948 3.58494 22.1313 4.51953 21.1113 5.11341C22.08 5.00235 23.0265 4.75012 23.9177 4.36518C23.2616 5.32141 22.4352 6.15435 21.479 6.82729Z\" fill=\"black\"></path>      </svg> </a></li>     <li><a href=\"https://www.linkedin.com/\" data-event=\"click\" data-category=\"social_link\" data-action=\"click\" data-section=\"footer\">     <svg width=\"24\" height=\"24\" viewbox=\"0 0 24 24\" fill=\"none\" xmlns=\"http://www.w3.org/2000/svg\">      <path d=\"M20.4498 20.4497H16.8936V14.8805C16.8936 13.5525 16.8699 11.843 15.0441 11.843C13.1919 11.843 12.9085 13.2898 12.9085 14.7838V20.4492H9.35238V8.99688H12.7662V10.5619H12.8141C13.5092 9.37346 14.8064 8.66098 16.1825 8.71207C19.7867 8.71207 20.4513 11.0828 20.4513 14.1671L20.4498 20.4497ZM5.33983 7.43138C5.33962 7.43138 5.3395 7.43138 5.3394 7.43138C4.20728 7.43138 3.27569 6.4999 3.27569 5.36778C3.27569 4.23567 4.20728 3.30408 5.3394 3.30408C6.47141 3.30408 7.40289 4.23545 7.4031 5.36736C7.4031 5.36746 7.4031 5.36757 7.4031 5.36778C7.4031 6.49969 6.47173 7.43128 5.33983 7.43138ZM7.11788 20.4497H3.55802V8.99688H7.11788V20.4497ZM22.2226 0.00165575H1.77096C0.810928 -0.00918665 0.0117012 0.771465 0 1.7315V22.2681C0.0112718 23.2287 0.810391 24.0102 1.77096 23.9999H22.2226C23.1856 24.012 23.9875 23.2306 24.0004 22.2681V1.7301C23.9875 0.768029 23.1845 -0.0125153 22.2226 0.000152033V0.00165575Z\" fill=\"black\"></path>      </svg> </a></li>    </ul>    </div>   </div>   </section>  </footer>  <!--=== CART ===-->  <section class=\"unstack__drawer unstack__drawer--left\" id=\"unstack__drawer_shopify\" data-drawer-target=\"\">   <div onclick=\"window._shop.toggleCart()\" class=\"unstack__drawer__overlay\" data-drawer-close=\"\" tabindex=\"-1\"></div>   <div class=\"unstack__drawer__wrapper\">   <div id=\"unstack__drawer_loader\" class=\"unstack__drawer_loader_wrapper\">    <span class=\"unstack__drawer_loader\"></span>   </div>   <div class=\"unstack__drawer__header\">    <div></div>    <div class=\"unstack__drawer__title\">   Your cart   </div>    <button onclick=\"window._shop.toggleCart()\" class=\"unstack__drawer__close\" data-drawer-close=\"\" aria-label=\"Close Drawer\"></button>   </div>   <div class=\"unstack__drawer__content\" id=\"unstack__drawer__content\">    <div data-cart-content=\"\"></div>   </div>   <!-- <span id=\"unstack__drawer__footer\" class=\"unstack__drawer__footer\">   <span id=\"unstack__drawer__error\" class=\"unstack__drawer_error\" style=\"display: none;\"></span>   <a id=\"unstack__cart__checkout\" href=\"#\" target=\"_blank\" class=\"button button-primary button-full\">Checkout</a>  </span> -->   </div>  </section>  <!--=== CART ===-->  <!-- Component --> <section data-id=\"e247708a-a788-4811-90dd-694358c3f785\" class=\"component-7e66e313-14ac-400b-a80d-b7ace502c3e0 section-e247708a-a788-4811-90dd-694358c3f785\">   <div id=\"gdprModal\" style=\"display: flex;\">   <div class=\"upperContent\">    <h3 class=\"title\">We value your privacy</h3>    <div class=\"message\">   We use cookies to customize your browsing experience, serve personalized ads or content, and analyze traffic to our site.   </div>   </div>   <div class=\"buttons\">    <button type=\"button\" class=\"button button-primary button button-primary\" onclick=\"window._site._widgets.Gdpr.accept()\">Accept all</button>    <button type=\"button\" class=\"button button-secondary button button-secondary\" onclick=\"window._site._widgets.Gdpr.reject()\">Strictly Necessary Only</button>   </div>   </div>  </section> <!-- /Component -->  <div class=\"templates\">     <!-- /Animation -->  </div> </body></html>";
+		String generalized_src2 = Browser.cleanSrc(src2);
+		System.out.println(generalized_src2);
+		assertTrue(expected_result2.equals(generalized_src2));
+	}
+}
